web_app/requirements.txt
torch~=1.0
torchvision
sklearn
pandas
pynvml
flask
Pillow
